9/29

+++ 電腦圖學導論/正課/Project1 FLTK/Skeleton/Codes/TargaImage.h
@@ -66,7 +66,7 @@ class TargaImage
         bool Half_Size();
         bool Double_Size();
         bool Resize(float scale);
-        bool Rotate(float angleDegrees);
+        bool Rotate(double angleDegrees);
 
     private:
 	// helper function for format conversion
